Alterações gerais e adição de Crud Model

+++ database/migrations/2018_11_03_022028_create_tags_table.php
@@ -14,7 +14,7 @@ class CreateTagsTable extends Migration {
 	{
 		Schema::create('tags', function(Blueprint $table)
 		{
-			$table->integer('id', true);
+            $table->bigInteger('id', true);
 			$table->string('texto', 80)->unique('texto_UNIQUE');
 			$table->timestamps();
 			$table->softDeletes();
